research_publisher/pyproject.toml
[project]
name = "research-publisher"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "crewai>=0.157.0",
    "crewai-tools>=0.0.1",
    "datasets>=4.0.0",
    "ddgs>=9.5.2",
    "duckduckgo-search>=8.1.1",
    "ipython>=9.2.0",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-ollama>=0.3.6",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.4.7",
    "linkup>=0.1.4",
    "mcp[cli]>=1.9.2",
    "nbconvert>=7.16.6",
    "nbformat>=5.10.4",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.48.0",
    "unstructured>=0.18.11",
]
